update readme.md
BREAKING CHANGE: `comment` properties (containing the full extract of each leading comment
(with all the `/**` and the ` * `) are entirely removed from the documentation as they were not useful.
 Use `description` or `tags` instead.

BREAKING CHANGE:  Events that are not commented near their call
will not anymore be parsed. If you want to document events,
do it as a leading comment before your event call